changes in alerts

+++ Frontend/static/js/metadataInfoscript.js
@@ -1,4 +1,23 @@
 function deleteButton(){	
 	ident = document.getElementById("identifier").value
-	window.location.href="/delete/"+ident
-};
+	$.confirm({
+		title: 'Alert!',
+		content: 'Deleting this dataset might take a little time. Do you want to delete it?',
+		type: 'red',
+		useBootstrap: true,
+		typeAnimated: true,
+		buttons: {
+			delete: {
+				text: 'Delete',
+				btnClass: 'btn-red',
+				action: function(){
+					return window.location.href="/delete/"+ident
+				}
+			},
+			cancel: function(){
+
+			}
+		}
+	});
+	
+}
